Alteração de Int para Long nos IDs

+++ src/main/java/br/com/soc/sistema/business/AgendaBusiness.java
@@ -73,7 +73,7 @@ public class AgendaBusiness {
 	public AgendaVo buscarAgendaPor(String codigo) {
 		
 		try {
-			Integer cod = Integer.parseInt(codigo);
+			Long cod = Long.parseLong(codigo);
 			return dao.findByCodigo(cod);
 		} catch (NumberFormatException e) {
 			throw new BusinessException(FOI_INFORMADO_CARACTER_NO_LUGAR_DE_UM_NUMERO);
