+++ src/main/java/br/com/soc/sistema/dao/FuncionarioDao.java
@@ -35,7 +35,7 @@ public class FuncionarioDao extends Dao {
 			
 			int i=1;
 			ps.setString(i++, funcionarioVo.getNome());
-			ps.setInt(i++, Integer.parseInt(funcionarioVo.getRowid()));
+			ps.setLong(i++, Long.parseLong(funcionarioVo.getRowid()));
 			ps.executeUpdate();
 			
 		} catch (SQLException e) {
@@ -50,7 +50,7 @@ public class FuncionarioDao extends Dao {
 			PreparedStatement ps = con.prepareStatement(query.toString())) {
 			
 			int i=1;
-			ps.setInt(i, Integer.parseInt(funcionarioVo.getRowid()));
+			ps.setLong(i, Long.parseLong(funcionarioVo.getRowid()));
 			ps.executeUpdate();
 			
 		} catch (SQLException e) {
@@ -111,7 +111,7 @@ public class FuncionarioDao extends Dao {
 		return Collections.emptyList();
 	}
 	
-	public FuncionarioVo findByCodigo(Integer codigo){
+	public FuncionarioVo findByCodigo(Long codigo){
 		StringBuilder query = new StringBuilder("SELECT rowid id, nm_funcionario nome FROM funcionario ")
 								.append("WHERE rowid = ?");
 		
@@ -119,7 +119,7 @@ public class FuncionarioDao extends Dao {
 			PreparedStatement ps = con.prepareStatement(query.toString())){
 			int i = 1;
 			
-			ps.setInt(i, codigo);
+			ps.setLong(i, codigo);
 			
 			try(ResultSet rs = ps.executeQuery()){
 				FuncionarioVo vo =  null;
